Crate pallet_revive_proc_macro

Procedural macros used in pallet-revive.

Most likely you should use the #[define_env] attribute macro which hides boilerplate of defining external environment for a polkavm module.

Attribute Macros

define_env

Defines a host functions set that can be imported by contract polkavm code.

unstable_hostfn