pallet_revive_proc_macro

Attribute Macro define_env

#[define_env]

Defines a host functions set that can be imported by contract polkavm code.

CAUTION: Be advised that all functions defined by this macro cause undefined behaviour inside the contract if the signature does not match.

WARNING: It is CRITICAL for contracts to make sure that the signatures match exactly. Failure to do so may result in undefined behavior, traps or security vulnerabilities inside the contract. The runtime itself is unharmed due to sandboxing. For example, if a function is called with an incorrect signature, it could lead to memory corruption or unexpected results within the contract.